requirements.txt
fastapi
uvicorn
razorpay
python-dotenv
pydantic
email-validator
requests